refactor: interview-config featureにrepositoryレイヤーを追加
