+++ web/src/features/interview-config/server/loaders/get-interview-config-admin.ts
@@ -1,7 +1,10 @@
 import type { Database } from "@mirai-gikai/supabase";
-import { createAdminClient } from "@mirai-gikai/supabase";
 import { unstable_cache } from "next/cache";
 import { CACHE_TAGS } from "@/lib/cache-tags";
+import {
+  findLatestInterviewConfigByBillId,
+  findPublicInterviewConfigByBillId,
+} from "../repositories/interview-config-repository";
 
 export type InterviewConfig =
   Database["public"]["Tables"]["interview_configs"]["Row"];
@@ -18,15 +21,9 @@ export async function getInterviewConfigAdmin(
 
 const _getCachedInterviewConfigAdmin = unstable_cache(
   async (billId: string): Promise<InterviewConfig | null> => {
-    const supabase = createAdminClient();
-
     // まず公開設定を探す
-    const { data: publicData, error: publicError } = await supabase
-      .from("interview_configs")
-      .select("*")
-      .eq("bill_id", billId)
-      .eq("status", "public")
-      .single();
+    const { data: publicData, error: publicError } =
+      await findPublicInterviewConfigByBillId(billId);
 
     if (publicData) {
       return publicData;
@@ -34,13 +31,8 @@ const _getCachedInterviewConfigAdmin = unstable_cache(
 
     // 公開設定がなければ、最新の更新日の設定を返す
     if (publicError?.code === "PGRST116") {
-      const { data: latestData, error: latestError } = await supabase
-        .from("interview_configs")
-        .select("*")
-        .eq("bill_id", billId)
-        .order("updated_at", { ascending: false })
-        .limit(1)
-        .single();
+      const { data: latestData, error: latestError } =
+        await findLatestInterviewConfigByBillId(billId);
 
       if (latestError) {
         if (latestError.code === "PGRST116") {
